Feat: Add Dojo code
This commit adds the code that was created in the dojo. Our group split
into two sub teams. We agreed (more or less!) a common interface for a
"Lift" class and one team worked on the display and the other team
worked on the logic. With minimal integration time, these the work was
pasted into the same module and while some of the logic is questionable,
it all worked together pretty well!